Current codebase is compatible with PHP 7.1
Because the code currently used works fine on PHP 7.1 there is no need to force people to have a higher version. Also this package replaces the old api-restclient which supported PHP ^5.6 so to make it easier for people to switch make the gap as small as possible. (ofcourse people should be encouraged to use the latest PHP version possible :wink:)